backend/requirements.txt
fastapi==0.109.2
python-multipart==0.0.9
httpx==0.26.0
python-dotenv==1.0.0
uvicorn==0.27.1